Basics

+++ rpgmv_translator/utils.py
@@ -5,6 +5,21 @@ import re
 # from nltk.tokenize import word_tokenize
 
 
+def read_progress_log(directory):
+    progress_log_path = os.path.join(directory, 'progress.log')
+    progress = {}
+    if os.path.exists(progress_log_path):
+        with open(progress_log_path, 'r') as file:
+            for line in file:
+                key, value = line.strip().split(':')
+                progress[key] = value == 'completed'
+    return progress
+
+def update_progress_log(directory, step):
+    progress_log_path = os.path.join(directory, 'progress.log')
+    with open(progress_log_path, 'a') as file:
+        file.write(f"{step}:completed\n")
+
 def is_rpgmv_folder(directory):
     required_folders = ['audio', 'data', 'img']
     www_path = os.path.join(directory, 'www')
@@ -25,11 +40,23 @@ def duplicate_json_files(directory):
             if filename.endswith('.json'):
                 original_path = os.path.join(dirpath, filename)
                 backup_path = os.path.join(dirpath, f"{filename}.old")
-                shutil.copy2(original_path, backup_path)
+
+                # Check if the backup file already exists
+                if not os.path.exists(backup_path):
+                    shutil.copy2(original_path, backup_path)
 
 def restore_from_backup(directory):
     if not is_rpgmv_folder(directory):
         raise ValueError("The specified directory is not a valid RPGMV folder.")
+
+    # List of new files that should be removed if they exist
+    new_files = ['original.csv', 'translated.csv', 'progress.log']
+
+    # Remove new files if they exist
+    for new_file in new_files:
+        new_file_path = os.path.join(directory, new_file)
+        if os.path.exists(new_file_path):
+            os.remove(new_file_path)
 
     restored_any = False
     for dirpath, dirnames, filenames in os.walk(directory):
@@ -38,6 +65,7 @@ def restore_from_backup(directory):
                 backup_path = os.path.join(dirpath, filename)
                 original_path = os.path.join(dirpath, filename.replace('.old', ''))
                 shutil.copy2(backup_path, original_path)
+                os.remove(backup_path)
                 restored_any = True
 
     if not restored_any:
